+++ include/votca/csg/potentialfunctions/potentialfunctionljg.h
@@ -31,9 +31,9 @@ class PotentialFunctionLJG : public PotentialFunction {
   // calculate function value for given r
   double CalculateF(const double r) const override;
   // calculate first derivative w.r.t. ith parameter
-  double CalculateDF(const int i, const double r) const override;
+  double CalculateDF(const long int i, const double r) const override;
   // calculate second derivative w.r.t. ith parameter
-  double CalculateD2F(const int i, const int j, const double r) const override;
+  double CalculateD2F(const long int i, const long int j, const double r) const override;
 };
 }  // namespace csg
 }  // namespace votca
