+++ src/tools/csg_dump.cc
@@ -81,7 +81,7 @@ bool CsgDumpApp::EvaluateTopology(Topology *top, Topology *top_ref) {
       cout << "molecule: " << (*mol)->getId() + 1 << " " << (*mol)->getName()
            << " beads: " << (*mol)->BeadCount() << endl;
       for (int i = 0; i < (*mol)->BeadCount(); ++i) {
-        int resnr = (*mol)->getBead(i)->getResnr();
+        long int resnr = (*mol)->getBead(i)->getResnr();
 
         cout << (*mol)->getBeadId(i) << " Name " << (*mol)->getBeadName(i)
              << " Type " << (*mol)->getBead(i)->getType() << " Mass "
